Pull out test fixture setup into a module.
